pyproject.toml
[project]
name = "capstone-project-kaban"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.127.0",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "psycopg2>=2.9.11",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

[tool.ruff.lint.per-file-ignores]
"app/db/create_db.py" = ["F401"]
"**/models/__init__.py" = ["F401"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:passlib.*"
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
